feat: file cleanup
