--- src/ui/views/strategies/export_strategy.rs
@@ -1,53 +1,0 @@
-//! Strategy export-to-file form and screen.
-
-use tuirealm::{event::KeyEvent, tui::prelude::Rect, Frame};
-
-use crate::{
-    backend::{StrategyTask, Task},
-    ui::form::{FormController, FormStatus, Input, InputStatus, SelectInput},
-};
-
-pub(super) struct ExportStrategyFormController {
-    input: SelectInput<String>,
-}
-
-impl ExportStrategyFormController {
-    pub(super) fn new(strategies: Vec<String>) -> Self {
-        Self {
-            input: SelectInput::new(strategies),
-        }
-    }
-}
-
-impl FormController for ExportStrategyFormController {
-    fn on_event(&mut self, event: KeyEvent) -> FormStatus {
-        match self.input.on_event(event) {
-            InputStatus::Done(strategy_name) => FormStatus::Done {
-                task: Task::Strategy(StrategyTask::ExportStrategy(strategy_name)),
-                block: false,
-            },
-            InputStatus::Exit => FormStatus::Exit,
-            status => status.into(),
-        }
-    }
-
-    fn form_name(&self) -> &'static str {
-        "Strategy export"
-    }
-
-    fn step_view(&mut self, frame: &mut Frame, area: Rect) {
-        self.input.view(frame, area)
-    }
-
-    fn step_name(&self) -> &'static str {
-        ""
-    }
-
-    fn step_index(&self) -> u8 {
-        0
-    }
-
-    fn steps_number(&self) -> u8 {
-        1
-    }
-}
